Added PCB files as well as Gerber
Added electronics files for fabrication and modification. Please take the board with a grain of salt as the intention was to have something working with minimal effort that's easy to repair and to reproduce if needed. I also wanted to be able to mill the board using my PCB mill, hence the large traces.